Safari-609.1.11.1/LayoutTests/crypto/subtle/hmac-generate-key-sha512-expected.txt

Test generating a HMAC key with sha-512.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Generating a key...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'HMAC'
PASS key.algorithm.length is 1024
PASS key.algorithm.hash.name is 'SHA-512'
PASS key.usages is ["sign", "verify"]
PASS successfullyParsed is true

TEST COMPLETE
